ComposeBox: Make actual button a Touchable, not its container
Now, a tight square around the circular button itself, not an
arbitrarily larger square around it, gives feedback when the button
is pressed. On Android, we need to mask the ripple effect so it
doesn't cover a square at all, but rather just the circle itself.
We'll do that next.

We don't want to shrink the touch target, so give the button some
hitSlop so the touchable area is the same.

In 5d24a3dbd, in 2018-11, we intentionally switched the Touchable
and the View to make the Touchable the parent, which we reverse
here. Reading the reasoning in that commit, it looks like a main
concern was keeping the touchable area large enough. I suspect we
would have used hitSlop instead of doing that, if we had known about
it at the time.