ChatScreen: If no unreads, use last message ID for initial scroll
This is basically planned followup to #3457 [1], with #4920 in mind,
but with a slightly different approach.

The proposal I made there was to change `getFirstUnreadIdInNarrow`
so it returns `messages[messages.length - 1]` if the first unread ID
doesn't exist, i.e., if there are no unread messages.

But that's not great; if we do that, then an ID returned by
`getFirstUnreadIdInNarrow` won't always be…the first unread ID in
the narrow. That function is right to return `null` when "the first
unread ID in the narrow" can't be given because it doesn't exist.

But after calling that function, we can conditionalize on a `null`
value so that we still use the last message ID for
`initialScrollMessageId`. So, do that.

The last message (`messages[messages.length - 1]`) still might not
exist because `messages` might be empty. In that case, carry on
passing `null`.

[1] https://github.com/zulip/zulip-mobile/issues/3457#issuecomment-753226145