android notif [nfc]: Further revise comment about PendingIntent
Mostly make it a bit shorter (deferring to the upstream doc for
details), and try to clarify the situation about two PendingIntents
being the same.

Also let the prose run a few columns further to the right.  This is
pretty deeply indented, so that going to column 90 still means it's
only 66 columns wide, which is a fine width for reading text.