android notif: Remove NotificationIntentService
And replace it with directly opening MainActivity.

When we set targetSdkVersion to 31, we will no longer be allowed to
use a Service as a "notification trampoline", for performance
reasons:
  https://developer.android.com/about/versions/12/behavior-changes-12#notification-trampolines

So, go with their recommended fix, and have the PendingIntent open
MainActivity directly.

In the preceding commits, we moved everything out of
NotificationIntentService except the code that starts MainActivity,
to make this commit as small as possible. (The really important bit
was the `notifyReact` call, which lets the JS environment know about
the notification the user opened.)

Fixes: #5145