storage tests [nfc]: Trim comments about AsyncStorage mocks
We're about to switch to our own AsyncStorage replacement.  That
will make many of the specifics in these comments no longer true --
it doesn't provide mocks for these, instead it runs the actual code
-- but the upshot for what we want to mock in these tests remains
the same.

So, shorten these comments to what we'll want them to be after that
change.  The resulting comments are all accurate already, too.

Similarly, give a less specific name to the locals where we save the
old values of these methods before mocking them for these tests.