TranslatedText [nfc]: Inline last caller and dissolve
There really hasn't been a use for the "without applying any
styling" part of TranslatedText's interface. Everywhere we'd ever
want to use `TranslatedText`, we'd want to apply to the text either

- Zulip's signature default styling (including for dark/light
  themes), or
- That default styling, plus some special styling for the callsite

This was true of TranslatedText's last caller, removed here. It was
also true of the two other callers we removed in the previous
commit.

But the point applies in general:

1. We're responsible for applying styles to text in the entire React
   component tree. Those styles must be computed with input from our
   app-specific default styles in every case.
2. `TranslatedText` is used in the React component tree and nowhere
   else. It can't be used anywhere else because it's a React
   component.
3. So we'd need to apply some styles (at least the default styles)
   everywhere `TranslatedText` could be used.

That's not to say there isn't a use for bare, styleless strings that
are translated. We create those as plain strings, using our `_`
wrapper around `intl.formatMessage`. We only *need* those plain
strings when a React element won't do; in particular, when we're
rendering text through a thinly wrapped native API that asks for a
string [1]. Like when we tell the iOS native action sheet what its
buttons should say.

[1] I think it's probably a bug to use the plain strings for
    anything else. `RawLabel`, which we'll soon rename to
    `ZulipText`, is where we define our app-specific default text
    styles. Those default styles are important: they ensure
    readability in dark and light mode, they help set a consistent
    look-and-feel, and they've been chosen to represent Zulip. So
    any code that renders text and doesn't clearly go through that
    component, but could, should be a code smell.