license: Bring THIRDPARTY file up to date
It's 2021, and we've kept making changes, so update the copyright
years for our own work.

We've also made a number of changes to the redux-persist-migrate
code.  Like all our software, we make that work available under
the Apache License, Version 2.0 ("Apache-2").  Because the MIT
license (which the upstream version of that code is provided
under) is compatible with Apache-2, the combination is available
under simply the same Apache-2 license.

There's also some more third-party code for which we forgot to
update this file: in src/third/redux-persist/, added as e5409c578
in 2020-05.

 * The upstream for that code has the MIT license, with a
   copyright notice "Copyright (c) 2015-present Zack Story".

 * The last upstream commit (github:rt2zz/redux-persist@v4.10.2)
   for that code is dated 2017, so "present" really means 2017.

 * Because again the MIT license is compatible with our Apache-2
   license, the new versions with our changes are under the same
   Apache-2 license as the rest of our code.

Finally, we deleted vendor/intl/, in 30e4d19cc in 2020-08.