migration tests: Test all existing migrations
Really we should have done this years ago: this is code
that it's important should behave correctly, and it doesn't
naturally get much manual testing from simply using the app.
Well, the second-best time is now.

In fact, writing these tests turns up a bug in one of the
old migrations!

We'll fix that bug in a separate commit.  Fortunately its
consequences aren't too bad -- it caused that migration to
drop PM drafts, which is exactly what we'd chosen to do in
the case of the migration right before it.  But it makes a
nice demonstration of why tests are necessary here.