narrow: Switch to stream ID at some caseNarrow call sites
This covers the spots where this conversion is local and trivial.

The change should be NFC except in cases where the stream name
and ID don't agree (e.g., because the stream's name has been
changed since we formed this Narrow object).  In those cases,
it fixes a bug.

To find these, and the other call sites in the next few commits,
I manually audited through all calls of the `caseNarrow*` functions.
Ideally for this sort of refactor one would like the type-checker to
do that work, but sadly here it isn't able to completely do so: in
getComposeInputPlaceholder we pass the callback's argument to string
interpolation, and in two other places we used it only for `===`.
(If there were a `JSON.stringify`, that'd have the same problem.)
And those accept numeric IDs just the same as string names.