deps: Force sqlite3 to use a recent node-gyp
In particular this leads to using a reasonably recent `tar` package,
fixing vulnerabilities in the old one it was using.

Upstream has already bumped this to node-gyp 7.x in their master
branch, but haven't posted a release to NPM:
  https://github.com/mapbox/node-sqlite3/issues/1493

Empirically node-gyp 8.x, the latest, works fine.  That's also
reported by someone on that issue thread:
  https://github.com/mapbox/node-sqlite3/issues/1493#issuecomment-905633454
May as well go for that, then.  (There was no 8.x yet when the
version specified in sqlite3 was bumped to 7.x.)

Some other people on that thread report using a fork made by the
VS Code developers, which posted some releases in November.  But
that fork seems pretty clearly intended for VS Code's own internal
use, with no promises for broader consumption:
  https://github.com/microsoft/vscode-node-sqlite3/issues/14#issuecomment-1009377469
so that doesn't seem like an improvement over upstream.