persist [nfc]: The REHYDRATE payload is not void
I'm not sure if this was ever a possible case in the past; it
wouldn't be surprising if it was something eliminated in some of
the work we did on redux-persist a couple of months ago.

In any case, it seems to clearly not be possible now: looking at
`persistStore.js` where we create these actions, the payload is
always an object.