notif types: Document APNs payload with actual type syntax
This gives this information more room to spread out than it has
embedded inside a comment.  In particular, this makes it possible
to use JSDoc syntax `/** … */`, like we do for other types; that
won't work inside a block comment, because JS comments don't nest.

Also move this information to src/api/, as like the rest of that
subtree it's describing part of the Zulip server API.