android build: Upgrade Gradle to 6.9, following Expo template app
Note: If you get an error with lots of lines like
  Entry name 'assets/fonts/AntDesign.ttf' collided
, try
  cd android && rm -rf .gradle/ && ./gradlew clean
.

Done now to follow the change to
templates/expo-template-bare-minimum in expo/expo@f05c8ccb6.

Done by following our instructions at the top of the changed file.

The second time I ran the command in those instructions, I got this
helpful output:

  Welcome to Gradle 6.9!

  Here are the highlights of this release:
   - This is a small backport release.
   - Java 16 can be used to compile when used with Java toolchains
   - Dynamic versions can be used within plugin declarations
   - Native support for Apple Silicon processors

  For more details see https://docs.gradle.org/6.9/release-notes.html

All of that looks nice and helpful.