reducer: Split out the initialization case
That is, split the case where the previous state is void, and so the
action must be the store initialization action, from the case where
it's any other action and the previous state must be of its usual
object type.

This makes for a bit more boilerplate, and one more line to update
when adding or removing a sub-reducer.  But both the split versions
are simpler, and that will help us add new complications that we need.