docs/architecture: Cut long-outdated discussion of code structure
A lot of this hasn't been true since 2017 or 2016.  We could bring it
up to date... but I'm not sure there's a lot of value in a description
like this, vs. just browsing the actual code.  It doesn't seem like
anyone's been reading it, or else we'd have noticed more how out of
date it is.