msglist: Drop `narrow` param from getEditSequence
Now that this isn't being passed to messageListElementHtml,
the only thing we're doing with it is forcing an update if
it's changed.

But because it isn't being passed to messageListElementHtml,
it's no longer possible that it affects the resulting HTML
and therefore should cause an update.  So we don't need to
do that either.

(A change in narrow still can cause things to change, but it'd
do so only by affecting the message-list elements themselves.)