webview js [nfc]: Clean up isDevelopment variable
We've had a version of this since 3645ffe47, in 2018.  Much more
recently, we worked out the pattern followed by platformOS and
doNotMarkMessagesAsRead: the variable is a local, rather than stuck
onto `window`; and rather than encode it with any ad-hoc logic, we
do so with `JSON.stringify`, which gives a single general way to
encode (as a valid fragment of JS code) any data we're likely to
want to pass through this channel.

Also fix the type declaration: this is a boolean.  One of the
benefits of using `JSON.stringify` here is that it becomes easier
to see that.