stream [nfc]: Fold out PseudoSubscription in SubscriptionsCard
Here, the items we're passing are always in fact Subscription values,
so we can just say that.

This means that the value we pass for the `isSubscribed` prop --
the value we always have been passing there, for the StreamList
inside SubscriptionsCard -- is actually `undefined`, so falsy,
even though in reality the user is subscribed to every one of these.

That's only a latent bug, though, because we have showSwitch false,
and in that case StreamItem doesn't look at isSubscribed.
Note that in a comment here, and jsdoc there.