message: Handle update_message in narrows/caughtUp, simply
This isn't optimal in that if, say, a conversation is moved, we'll
have to refetch the messages in the new topic to discover that it
includes the messages that were just moved.

But it is *correct* -- we won't mistakenly think we know something
we actually don't, and in particular won't get stuck in a state of
such wrong information.  And if the user looks at that conversation,
or is already looking, then we will indeed make that fetch.

The reason keeping things correct is tricky is that the target
stream and topic may already have some messages -- in fact this is
a fairly common use case.  We may not know anything about that
narrow, and so we may not be able to tell if that's the case.
Leave some TODO comments about ways we could use knowledge we have
when applicable, though.

Fixes: #2688