message: Handle update_message making stream/topic changes
This handles part of #3408, closely related to #2688.  In particular
this means that if you find one of these messages in an interleaved
view where recipient headers are shown, the recipient header will
take you to the correct narrow.

We actually did handle topic edits, but only for a single message,
and didn't handle moves between streams.