deps: Upgrade react-native-vector-icons to 9.1.0, the latest
This release reportedly fixes the issue that made the "@" icon look
weird (the issue that was keeping us below 7.1.0):
  https://github.com/oblador/react-native-vector-icons/pull/1311#issuecomment-1038258615
I've checked on iOS and Android, and the icon looks right at this
version.

See changelog:
  https://github.com/oblador/react-native-vector-icons/releases

A few breaking changes were announced, but none that would affect
us:

- Icon.TabBarItemIOS convenience component removed as it was removed
  from React Native core.
- Breaking change: Drop support for
  react-native-web/react-native-desktop aliasing.

I didn't notice anything wrong with the icons on a quick tour of the
app at this version, on iOS or Android.