notif [nfc]: Split out module for NotificationListener
The way we've had much of our notifications code split, where there's
one module for "actions" and then the rest goes in a general "index"
module, isn't really all that helpful.  Signs of that include:

 * Each module imports several definitions from the other.  When
   reading code in either file, one typically ends up jumping back and
   forth from one file to the other in order to understand what's
   happening.  So the distinction isn't successfully drawing a line
   between meaningful subsystems.

 * Several of the non-"action" functions end up looking awfully
   action-like anyway, taking a "dispatch" argument.  One,
   `androidGetToken`, even uses its "dispatch" to get a "getState"
   function too, by dispatching an ad-hoc action.  So it's not clear
   the distinction between "action" code and otherwise is even
   meaningful here at all.

Instead, we'll reorganize this code to reflect the two largely-
separate tasks it performs: setting up for the server to be able to
send this device notifications in the first place, and acting on a
notification when the user opens it.

There's one chunk of this code that straddles those two as a caller:
where we set up to listen for notification-related events.  Start by
moving that to its own module.