UserStatusScreen: Rename "Update" button to "Save"; remove checkmark icon
I think this is much more common wording.

And the checkmark icon is potentially confusing because it looks
just like the checkmark on one of the status suggestions
(SelectableOptionRow), if you've selected one. And it just doesn't
feel normal to have an icon at all for a simple "OK" button like
this.