api [nfc]: Put params-object last in updateMessage
This feels a bit more conventional: the grab-bag of what amount to
optional, named, parameters comes after all the positional parameters,
which are required.

It also produces much nicer formatting from Prettier at call sites
if the params-object gets long: the other arguments get to stay on
the first line.