SharingScreen: Use `resetToAccountPicker`, not `navigateToAccountPicker`
There's no reason to keep a SharingScreen in the nav stack in this
case. There won't be anything useful on it -- it depends on a
logged-in account with server data -- and the only way to change
that is by following the auth flow, which will result in a
`resetToMainTabs` that clears the SharingScreen off the stack
anyway.