vscode: Specify what auto-formatter to use
This is needed in order for format-on-save to work; without it,
VS Code doesn't know whether to use this or its built-in TS/JS
formatter.

I'd had the same setting in my personal VS Code settings on my
main desktop, so hadn't noticed its absence.  But really it's a
project-specific fact that we use Prettier, so the project's
settings is where it belongs.