events [nfc]: Narrow scope of doInitialFetch and rename
This thunk action includes registering an event queue, of course --
but it has also included some unrelated tasks: notification setup
and an invocation of the broken outbox subsystem. Make this action
only about registering a queue and starting a polling loop on it,
pulling out that other stuff.

Also change its name to explicitly say what it does: not just the
"initial fetch", which I think basically means POST /register [1],
but also starting a polling loop, which should always be done
synchronously after that.

The fetch-PMs call stays in the function just because it's meant to
fill in a shortcoming of the register-queue functionality on older
servers. We still plan to remove that call when we desupport server
2.1.

Possibly the unrelated stuff should be grouped into a new thunk
action of its own, but I don't yet see a nice name or description
for such a group.

Next, we'll move this and its helpers to eventActions.js.

[1] See
      https://zulip.readthedocs.io/en/latest/subsystems/events-system.html#the-initial-data-fetch
    . But also "initial" can be a bit misleading, because it could
    misleadingly imply, for example, that this is only meant to run
    on startup and on first-time connection with an account. We run
    this with DEAD_QUEUE and ACCOUNT_SWITCH too, and will probably
    run it on `restart` events for #4793.