android build: Split up APKs per ABI
The great bulk of the size of our APK comes from the native
libraries in it (most of all libjsc.so).  Those exist in a
separate copy for each of our 4 ABIs, but any given device
is going to use only one of those.

So, build a separate APK for each ABI.  The new APK for armeabi-v7a
(the most widely-supported) is 35.1MB, compared to 132.3MB for the
old all-in-one APK.

This has no effect for the many Android users who get our app via
the Google Play Store: there, we already upload an AAB, which
Google splits up into several APKs still more finely.  See #3547.

Fixes: #5295