android build: Cut obsolete version-code-per-ABI hack
This bit of code doesn't currently do anything, because we build a
single APK for all devices.  It would if we started building split
APKs per ABI.

This sort of thing was required for publishing an app as multiple APKs
on Google Play.  (The docs link in the comment is broken; a current
version appears to be:
  https://developer.android.com/google/play/publishing/multiple-apks#VersionCodes
.)  Google Play would use the different version codes as tiebreakers
to decide which APK to give to a device when several were supported.

But on Google Play, publishing multiple APKs has been obsolete for
years.  Instead, one uploads a single AAB, "Android App Bundle", and
then Google slices and dices that into separate APKs for the per-ABI
files, the per-screen-density files, the per-screen-size files, and
so on.  Each device downloads the specific combination of those pieces
that applies to it.

We never did publish multiple APKs on Google Play, and because it's
now obsolete we never will; we've been using AABs since #3547 in 2020.

We do intend to start publishing multiple APKs, which would cause this
bit of code to start having an effect.  But this convention doesn't
seem to be useful anywhere else: for a user installing directly from
our GitHub release it's invisible and unhelpful, and the one known
downstream distributor of our APKs plans to just distribute the one
most widely-supported of them anyway:
  https://github.com/zulip/zulip-mobile/issues/1270#issuecomment-949969030
which will make this hack have no use there either.

So it seems like messing with the version codes like this would only
cause confusion.  Instead, just let the different per-ABI APKs for a
given release have the same version code as each other, and as the AAB
does.  That way the next release after v27.181 will have version code
simply 182, rather than 182003 et al.