example data: Add explicit full_name param to makeUser
The existing "name" parameter is confusing -- it's really meant for
a small convenience as used in selfUser, otherUser, and thirdUser,
but it looks so much like a synonym of `full_name` that we've ended
up using it in some tests where we really meant that.

Start offering an explicit `full_name` parameter, which we'll start
using instead.