react-intl libdef: Replace TS builtin Intl.NumberFormatPart Flow lacks
Well, sort of; see the comment, repeated at each site.

It'd be better if Flow gave an error at each of these sites, which
we could then suppress. Then when the Flow issue is fixed upstream,
we'd get an alert that the suppressions were no longer necessary,
and adapting would be as simple as removing the suppressions.