android build: Remove jcenter (mostly), following RN
Done to follow the template-app change in
facebook/react-native@9a923be89, on the path to the RN v0.65
upgrade.

We've already added lines for mavenCentral, following Expo, in
aedc8b87c. Here, follow RN in -- well, not quite removing the
jcenter lines, but narrowing their scope to a single dependency,
with a comment explaining why.

For why some lines are added to android/app/build.gradle, see the
commit message of the RN commit:

> Current flipper depends on Stetho version that is not available on
> MavenCentral, so had to exclude and bump the version.

A later commit, facebook/react-native@d8b115afa, removes those lines
and bumps Flipper to v0.93. We'll assume that's because Flipper
v0.93 is sufficiently not-"current" to allow removing the lines, and
we'll remove them too when we bump Flipper to v0.93.