emoji [nfc]: Use memoized selector for image emoji.
Introduces a new selector, getActiveImageEmoji which returns a list of
active image emojis in the correct shape to pass to @zulip/shared.

Also removes the construction of new objects for every matching emoji at
the return of getFilteredEmojis, as there is now no difference in the
shape of emoji objects.