streamsReducer tests [nfc]: Factor out a mkAction helper
When writing this helper, if we leave out an annotation on its
parameter `event`, then we end up with a bunch of Flow errors:

 * The type-checker sees that `event` can be any of several different
   shapes, from the different call sites.

 * It assumes that the type of `event` should be some one type -- just
   as it normally assumes that any given parameter, variable, etc.,
   should be some one type.  That type must therefore be a supertype
   of all of those types from the call sites -- so, a supertype of
   the union of those.

 * Then that doesn't work (when the result of `mkAction` is eventually
   fed to `streamsReducer`, which expects an `Action`), because there
   is no branch of `Action` where that union type can go.

   In particular, there's no branch of the `event` property on
   `EventAction` that's both able to have a `streams` property and
   able to have a `name` property, each of which appears at some of
   the call sites of `mkAction`.  This is what the error message
   focuses on:

       Cannot call streamsReducer with action bound to action because
       in property event: [incompatible-call]
        • Either property name is missing in object type [1] but
          exists in object literal [2].
        • Or property name is missing in object type [3] but exists
          in object literal [2].
        […]
        • Or property streams is missing in object type [6] but
          exists in object literal [2].

The solution is to tell Flow that the type of `event` won't actually
be any single type, but may be a different type at different call
sites: in other words, that `mkAction` is generic.

This is something that the type-checker won't reach for on its own,
because to have it try to guess where such a generalization will
actually help would mean giving it a much harder problem to solve --
would edge toward making it a full-blown theorem prover, rather than
a proof checker.  So when we do need it to generalize in that
direction, we have to give it a hint.

In particular, the types only work out because there is *some*
commonality between the call sites, some commonality which is
relevant in making sense of the definition.  So we have to identify
what that commonality is and provide it as a bound for the generic.
That's this:

    const mkAction = <E: $Diff<StreamEvent, {| id: mixed, type: mixed |}>>(event: E) =>

which says that when checking the body of `mkAction`, Flow should use
the assumption that `event` is a value (of some subtype `E`) of the
type `$Diff<StreamEvent, {| id: mixed, type: mixed |}>`.