StreamSettingsScreen: Make it a bit clearer what the "Edit" button does
And leave a TODO for something better.

The screen currently gives a jumbled collection of

- attributes about the stream and a button to change them

- a subscribe/unsubscribe button, and, if the user is subscribed,
  details about the subscription and switches to change the details

- a way to change who else is subscribed (but only by adding
  subscribers, and with no way to see the state before or after a
  change)

Each of these bullet points should be given its own coherent section
in the UI. Probably we should experiment with react-native-paper's
`Card`, as mentioned in the comment. This would be similar to what
Slack on iOS does.